moved items to single module file in base game engine directory to alleviate relative import headaches
