--- game_engine/item/key.py
@@ -1,12 +1,0 @@
-"""
-A key class. inherits from the item class.
-no parameters are specified
-"""
-from . import item
-
-class Key(item.Item):
-    """
-    The class used to specify a key object.
-    The key object has no actions.
-    """
-    pass
